change server name.
